Fix shadow and margin on DropDown
The class for the shadow is defined in tailwind, not DaisyUI.  The
DaisyUI classes have the dy- prefix (from tailwind.config.cjs). So the
class for shadow should not have the dy- prefix.

Include a top margin so that the dropdown does not sit on top of the app
bar.  If this DropDown is used somwhere else than the app bar, then we
will have to somehow pass in the margin or apply it from the outside.