Book navigation (#100)
* Book Navigation

Created a "SelectList" similar to SelectGrid.

Got relatively far, hit a roadblock with the "Book, Chapter, and Verse" tabs at the top. They're in the same list format as the book collections.

* Update BookSelector.svelte

* SelectList created

Unable to find the height parameter. Still need to do more research and find the component that affects that. Between Span and td

* Updated SelectList.svelte

Added more CSS, changed button to menu. Added a highlight feature. Added pointer cursor as well.

* Fixed Icon color

* Updated changes

* Committing changes since pr#102

* Updated Book Navigation

Bugged with on click

* Changes after PR#102

* Changed abbreviation to book name

* Use JavaScript style for parameter to bookGridGroup

When I was working with Jason, I suggested a style of parameters that
is more like other languages.  This is more idiomatic JavaScript. Also,
Allow the caller to specify which property to use as the label (better
design).

* Fix shadow and margin on DropDown

The class for the shadow is defined in tailwind, not DaisyUI.  The
DaisyUI classes have the dy- prefix (from tailwind.config.cjs). So the
class for shadow should not have the dy- prefix.

Include a top margin so that the dropdown does not sit on top of the app
bar.  If this DropDown is used somwhere else than the app bar, then we
will have to somehow pass in the margin or apply it from the outside.

* Remove conditional for selectionGrid in selectionList branch of code

---------

Co-authored-by: Chris Hubbard <chris@thehubbards.org>