Split body style into #content and #container styles
The styles generated are for the body element of the mobile
browser window in the Android app.  This browser window
doesn't include the app bar or tool bars.  With the PWA, we
need to split up the styles assigned to body. Most of the
styles will be attached to #container. The margin-top and
margin-bottom are changed to padding-top and padding-bottom
and attached to the #content element instead.