Feature annotation menus (#359)
* Test menu delete

This implements the delete function of the bookmarks menu. The function successfully deletes the bookmark, but it doesn't successfully update the cards on the page. Manually reloading the page does successfully update

* Enable deleting annotations (#4)

* Enable deleting annotations

* change display of annotations to be data-driven on sort order
* change load of annotations to mark depends and invalidate on delete

* Add compareColor

* Implement Annotation Menus

Implemented the delete, share, and view functions of the annotation menus. Also added functionality to navigate to verse when either annotation text or reference are clicked.

* Update Notes

Fixed issue with notes where docSet was not being initialized properly

* fix lint

---------

Co-authored-by: davidbbryant <123591459+davidbbryant@users.noreply.github.com>