Get numeral system from config and use to render
In `src/lib/scripts/configUtils.ts`, the getStyle function
gets a style attribute from the app's configuration, looking
first in the current book and then in the book collection.

In `src/lib/components/ScriptureViewSofria.svelte`, this function is
used to find a `numeralSystem` attribute. This attribute is then passed
to formatting functions to display chapter and verse numbers.