Fix c-drop direction to be based on current direction
In the native app, the sab-stylesheet is dynamically regenerated all
the time.  For the PWA, it is statically generated.

The c-drop float value needs to depend on the current direction of the
text.

I found this issue while viewing an app that was defined as RTL but
it had a book collection that was LTR.