Detect System Language of Browser (#519)
* Use browser language for UI if available

* Allow disabling system language detection

If the app is configured to always use the same language on first
startup, it will not check the browser for the system language.

* Allow use of secondary system language.

Browsers can provide a ranked list of preferred languages. Instead of
only checking the first preference, the app will now choose a default
language among all the preferred languages. If none of the preferred
languages are compatable with the app, the app will revert to the
fallback language.

* Set languageDefault when only one language is available

Previously, this variable was left undefined if the user had only one
choice of language.

* Change languageDefault to alias config.translationMapping.defaultLang