Feature/share app badges (#545)
* Include package name in config

Creating a link to the google play store requires the package name for the app.

* Display google play and apple store badges

If share-app-link or share-apple-app-link are enabled,
then display badges for the apppriate store.

The apple app store badge is not currently localized.

* Use localized appStore badges

Apple does not provide public links to app appStore badges for different
languages, so we included localized svg's in the project. Only copy the
languages used in the project data.

Only including en, fr at the moment. Will add other languages.

* Removed some console logs

* Added new badges for localizations.

Languages included are 'de' and 'es'. Spanish isn't working properly yet

* Moved the conditionals outside of the div

The reason for this is because it allows individual badges to be
centered even if the other badge's visibility is disabled.

* Fixed an issue with the spanish apple store badge not displaying.

Had to get the convertBadge.ts file to copy it into the static/badges
folder.

* Added more localized language options for the apple store badge.

The language codes added were AR, AZ, ID, KO, MAL, NL, PT, RU, TH, TR,
ZH. MAL needs to be reviewed to check if it is proper malaysian.

* Reorder tasks to prevent merge conflict

* format for lint

* Removed some unneeded comments

* format for lint

* Clean up from review

* Fix crash in convert when directory doesn't exist

* update package.json in another PR to keep from merge conflict

* format for lint Removed some unneeded comments. Added a needed one.

* Added href around buttons for share feature.

* Changed the href link for sharing apps so it goes to the google play or
apple store.

* Link the badges to appropriate stores

* correct lint error

---------

Co-authored-by: Chris Hubbard <chris@thehubbards.org>