Refactor reference store
The catalog is now an attribute of the reference store. Before use, the
reference store must be explicity intialized, allowing it to fetch the
catalog for the current docSet.

To simplify the reference store, much of its logic was refactored
into a class named `NavigationContext` in src/lib/navigation.ts.

BUG: In the browser, `NavigationContext` exhibits some bizzar behavior
while being initialized (each line is executed 6000+ times, leaf
functions are being called recursively, private variables are not getting
intitialized, etc.).