Catalog as Data (#532)
* Split catalog into one file per docSet

* Exclude catalog files from VSCode IntelliSense

VS Code extension was complaining about too many JavaScript
source files.

* Put catalog for current docSet into a store

TODO: The current implementation has a circular dependency between
stores:

The catalog store is derived from docSet, which is derived from refs,
which uses the catalog for the following:
- validate book and chapter
- calculate the number of verses in a chapter
- find the name/title of the book
- find the next/previous reference

This circular dependency will need to be resolved in a future commit.

* Refactor reference store

The catalog is now an attribute of the reference store. Before use, the
reference store must be explicity intialized, allowing it to fetch the
catalog for the current docSet.

To simplify the reference store, much of its logic was refactored
into a class named `NavigationContext` in src/lib/navigation.ts.

BUG: In the browser, `NavigationContext` exhibits some bizzar behavior
while being initialized (each line is executed 6000+ times, leaf
functions are being called recursively, private variables are not getting
intitialized, etc.).

* Rename catalog.ts to avoid naming conflict

The GitHub post-commit workflow creates a file `catalog.js`, which
conflicts with the new source file named `catalog.ts`. To avoid this
conflict, `catalog.ts` has been renamed to `catalogData.ts`.

* Fix bugs in reference store

* Remove unecessary option from jsconfig.json

There is no need to exclude the directory `static/collections` from
IntelliSense because it does not contain source files.

* Return to last reference on reload

* Use empty array for uninitialized fontChoices store