quiz squash #3
Modified the layout and page svelte files for the quizzes so they
work together

fix the quiz JSON file so the answers were correct for one
question, and the answers didn't show pictured, which were improper.

Made a layout.js file "+layout.js"

Started a layout svelte page for layout of quiz GUI. May not be needed.

recommitted the original changes for a new branch

Quizzes can now be viewed from dropdown menu in English.

Added a page.js file for quizzes, created a new id and collections
folder in the routes/quizzes folder and moved everything into that.

You can now navigate to the quizzes page by clicking on it in the
dropdown.

Rerouted navigation to the quiz page to include a docset folder.

Moved everything in the quiz collections folder into a docset folder

Moved this into the quizzes docset folder

A bunch of files got removed, so I had to rebuild PWA data files

the +page.js file for the quiz page now fetches quiz.json

Shifted some of the files around under the route/quiz folder.
Also changed the path in BookSelector to better match the path in the
quiz folder

Changed the path to the quiz files again.

in BookSelector, line 58: changed path from 'collection' to 'id'

fixed navigation to quiz

Removed unnecessary import