Remove workspaces
Makes the convert and data-sandbox packages indepenent from the root
package.json. This guarantees their dependencies will not be bundled
with the PWA.

The initial attempt to put convert and data-sandbox into workspaces
was unsuccessful because it required the src/ folder to also be in a
workspace. Specifically, SvelteKit would not create the necessary
`.svelte-kit/` directory on installation unless it was part of a
workspace.

Dependencies for data-sandbox will now have to be installed in its
own directory: `cd data-sandbox && npm i`