quiz squash #2
Deleted unnecessary layout files

format for lint

Added more elements to page.svelte

changes load data to be export instead of import

Fixed the path for navigating the static quiz folder in BookSelector

Corrected a mistake with changing the route for navigating to quizzes

Was able to get a quiz question with answers to pop onto the screen!

Removed languages from imports for quiz page. Not needed.

Formatted the quiz buttons in tailwind css and made them buttons.

got ready of unneeded comments

got rid of unneeded comments

Corrected the pathing for quizzes so that it now fetches a quiz from its
source.

Export quiz types

Got quizzes to cycle through questions and images to appear when they
exist.

Had to add a variable so images can be displayed on the quizzes page for
answers too.

Removed duplicate sound files in static

removed the appdef.xml file duplicate in static

format for lint

Playing around with the grid functionality for quiz answer buttons