Issue #900 - Plan doesn't work with verse range (#901)
If the plan segment end verse references a verse that is included as a combination verse (with something like data-verse="3-7"), the end of segment block is not included.  If trying to go to a reference that includes this kind of verse, the scroll to make it visible can't find it.